+++ tests/pvs1/test_auto_pvs1.py
@@ -36,8 +36,7 @@ def mock_strucvar():
 def mock_seqvar_pvs1(monkeypatch):
     mock_pvs1 = Mock(SeqVarPVS1)
     mock_pvs1.initialize.return_value = None
-    mock_pvs1.verify_PVS1.return_value = None
-    mock_pvs1.get_prediction.return_value = (PVS1Prediction.PVS1, PVS1PredictionSeqVarPath.NF1)
+    mock_pvs1.verify_PVS1.return_value = (PVS1Prediction.PVS1, PVS1PredictionSeqVarPath.NF1)
     monkeypatch.setattr("src.pvs1.auto_pvs1.SeqVarPVS1", lambda *args, **kwargs: mock_pvs1)
     return mock_pvs1
 
@@ -46,8 +45,7 @@ def mock_seqvar_pvs1(monkeypatch):
 def mock_seqvar_pvs1_failure(monkeypatch):
     mock_pvs1 = Mock(SeqVarPVS1)
     mock_pvs1.initialize.return_value = None
-    mock_pvs1.verify_PVS1.return_value = None
-    mock_pvs1.get_prediction.return_value = (None, None)
+    mock_pvs1.verify_PVS1.return_value = (None, None)
     monkeypatch.setattr("src.pvs1.auto_pvs1.SeqVarPVS1", lambda *args, **kwargs: mock_pvs1)
     return mock_pvs1
 
@@ -79,7 +77,6 @@ def test_autoPVS1_predict_seqvar_success(mock_seqvar_pvs1, mock_seqvar):
         autoPVS1.predict()
     assert mock_seqvar_pvs1.initialize.called
     assert mock_seqvar_pvs1.verify_PVS1.called
-    assert mock_seqvar_pvs1.get_prediction.called
     assert autoPVS1.seqvar_prediction == PVS1Prediction.PVS1
     assert autoPVS1.seqvar_prediction_path == PVS1PredictionSeqVarPath.NF1
 
@@ -91,7 +88,6 @@ def test_autoPVS1_predict_seqvar_failure(mock_seqvar_pvs1_failure, mock_seqvar):
         autoPVS1.predict()
     assert mock_seqvar_pvs1_failure.initialize.called
     assert mock_seqvar_pvs1_failure.verify_PVS1.called
-    assert mock_seqvar_pvs1_failure.get_prediction.called
     assert autoPVS1.seqvar_prediction == None
     assert autoPVS1.seqvar_prediction_path == None
 
@@ -103,7 +99,6 @@ def test_autoPVS1_predict_strucvar_success(mock_strucvar_pvs1, mock_strucvar):
         autoPVS1.predict()
     assert mock_strucvar_pvs1.initialize.called
     assert mock_strucvar_pvs1.verify_PVS1.called
-    assert mock_strucvar_pvs1.get_prediction.called
     assert autoPVS1.strucvar_prediction == PVS1Prediction.PVS1
     assert autoPVS1.strucvar_prediction_path == PVS1PredictionStrucVarPath.DEL1
 
@@ -115,6 +110,5 @@ def test_autoPVS1_predict_strucvar_failure(mock_strucvar_pvs1_failure, mock_stru
         autoPVS1.predict()
     assert mock_strucvar_pvs1_failure.initialize.called
     assert mock_strucvar_pvs1_failure.verify_PVS1.called
-    assert mock_strucvar_pvs1_failure.get_prediction.called
     assert autoPVS1.strucvar_prediction == None
     assert autoPVS1.strucvar_prediction_path == None
